Split TeamController methods from DevNestController
